+++ lib/src/list_tile.dart
@@ -38,6 +38,7 @@ class CustomListTile extends StatelessWidget {
   /// Dismissible wraps the [ListTile] inside a [Dismissible].
   CustomListTile.dismissible({
     super.key,
+    // TODO: this should only require ONE key
     required Object uniqueKey,
     required VoidCallback onDismissed,
     Color color = Colors.red,